ci(tests): decouple unit and integration tests
Unit and integration tests no run on separate commands

#18
