+++ docs/api/classes/client_LodestoneClient.default.md
@@ -28,4 +28,4 @@ Client for interfacing with the Final Fantasy XIV Lodestone.
 
 #### Defined in
 
-[client/LodestoneClient.ts:50](https://github.com/XIVStats/lodestone/blob/0cc3d3e/src/client/LodestoneClient.ts#L50)
+[client/LodestoneClient.ts:50](https://github.com/XIVStats/lodestone/blob/1f47efc/src/client/LodestoneClient.ts#L50)
