+++ jest.config.js
@@ -24,9 +24,10 @@
  */
 
 module.exports = {
-	preset: 'ts-jest',
-	transform: {
-		'^.+\\.(ts|tsx)?$': 'ts-jest',
-		"^.+\\.(js|jsx)$": "babel-jest",
-	}
-};
+  preset: 'ts-jest',
+  testMatch: ['**/__tests__/**/*.[jt]s?(x)', '**/__itests__/**/*.[jt]s?(x)', '**/?(*.)+(spec|test|itest).[jt]s?(x)'],
+  transform: {
+    '^.+\\.(ts|tsx)?$': 'ts-jest',
+    '^.+\\.(js|jsx)$': 'babel-jest',
+  },
+}
