+++ package.json
@@ -10,10 +10,16 @@
 		"build": "tsc",
 		"commit": "cz",
 		"docs": "typedoc --out docs/api --excludeNotDocumented --hideBreadcrumbs --readme none  src/**/*.ts",
-		"test": "jest",
-		"test:ci": "jest --ci --reporters=default --reporters=jest-junit --collect-coverage",
-		"format": "prettier --write src/**/*.ts src/**/*.test.ts",
-		"format:check": "prettier --check src/**/*.ts src/**/*.test.ts",
+		"test": "npm run test:unit && npm run test:integration",
+		"test:unit": "jest \"src\\/.*\\/__tests__\\/.*\\.test\\.ts\"",
+		"test:integration": "jest \"src\\/.*\\/__itests__\\/.*\\.itest\\.ts\"",
+		"test:u": "npm run test:unit",
+		"test:i": "npm run test:integration",
+		"test:ci": "npm run test:ci:unit && npm run test:ci:integration",
+		"test:ci:unit": "JEST_JUNIT_OUTPUT_FILE=reports/jest-unit.xml jest  \"src\\/.*\\/__tests__\\/.*\\.test\\.ts\" --ci --reporters=default --reporters=jest-junit --collect-coverage ",
+		"test:ci:integration": "JEST_JUNIT_OUTPUT_FILE=reports/jest-integration.xml jest \"src\\/.*\\/__itests__\\/.*\\.itest\\.ts\" --ci --reporters=default --reporters=jest-junit ",
+		"format": "prettier --write src/**/*.ts src/**/*.test.ts src/**/*.itest.ts",
+		"format:check": "prettier --check src/**/*.ts src/**/*.test.ts src/**/*.itest.ts",
 		"lint": "eslint --ext .ts --ext .js src --color",
 		"lint:fix": "eslint --fix --ext .ts --ext .js src --color",
 		"prepare": "husky install",
